SLING-8666 - slingstart-maven-plugin 1.8.4 or newer fail with "Failed to delete: .../target/features/processed: File does not exist: .../target/features/processed
Check that the processed directory actually exists.

+++ src/main/java/org/apache/sling/maven/slingstart/DependencyLifecycleParticipant.java
@@ -95,7 +95,9 @@ public class DependencyLifecycleParticipant extends AbstractMavenLifecyclePartic
                     env.modelProjects.put(project.getGroupId() + ":" + project.getArtifactId(), info);
                     File processed = new File(project.getBuild().getDirectory(), "features/processed");
                     try {
-                        FileUtils.forceDelete(processed);
+                        if ( processed.exists() ) {
+                            FileUtils.forceDelete(processed);
+                        }
                     } catch (IOException e) {
                         throw new MavenExecutionException("Failed to delete: " + processed.getPath(), e);
                     }
